+++ packages/@dopt/javascript-common/build.config.ts
@@ -1,0 +1,3 @@
+import { external } from '@dopt/pkg-build';
+
+export default external.node({});