+++ apps/@examples/card-custom/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @examples/card-custom
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [e8e8049bb]
+  - @dopt/react-rich-text@5.2.0
+  - @dopt/react-card@5.2.0
+  - @dopt/react@3.2.2
+  - @dopt/react-users@0.1.1
+
 ## 0.0.5
 
 ### Patch Changes