+++ apps/@examples/tour/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @examples/tour
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [e8e8049bb]
+  - @dopt/react-tour@6.2.0
+  - @dopt/react@3.2.2
+  - @dopt/react-users@0.1.1
+
 ## 0.1.13
 
 ### Patch Changes